+++ back/src/main/java/com/service/TokenService.java
@@ -12,6 +12,8 @@ import com.utils.PageUtils;
 
 /**
  * token
+ * @author yangliyuan
+ * @date 2019年10月10日 上午9:18:20
  */
 public interface TokenService extends IService<TokenEntity> {
  	PageUtils queryPage(Map<String, Object> params);
@@ -20,7 +22,7 @@ public interface TokenService extends IService<TokenEntity> {
    	
    	PageUtils queryPage(Map<String, Object> params,Wrapper<TokenEntity> wrapper);
 	
-   	String generateToken(Long userid,String username,String tableName, String role);
+   	String generateToken(Integer userid,String username,String tableName, String role);
    	
    	TokenEntity getTokenEntity(String token);
 }
